requirements.txt
flask==3.0.2
pillow==10.2.0
numpy==1.26.4
qiskit==1.1.0
matplotlib==3.8.3
Pillow>=8.3.1
opencv-python>=4.5.3.56
pydicom>=2.2.2
qiskit-aer>=0.12.0
pennylane>=0.24.0
scikit-learn>=0.24.2
scikit-image>=0.18.3
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.26.0 